Ignore InvalidAccessError.
